FreeType and Direct2d
Started on the basic principles of font rendering using the FreeType library, The system uses the FreeType cache manager which provides management of multiple fonts, and sizes. The drawText function includes that basic functionality for greyscale anti aliasing. The function also provides kerning adjusted faces. The font code is currently for windows only and not buffering to xcb on Linux yet,