Font Rendering and Layout
Many issues still exist within the layout manager. There are a few enhancements that correct layout issues. The latest demonstration shows wrapping and dynamic layout. The maximum of the heights is gathered at the element level. As well the layout is established by the recursive pen. However there are issues with the pen in the nested depths being reflected backwards. 

The problems of height is corrected by using the ascent to gather the baseline of the font.